backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend RAG System for Physical AI & Humanoid Robotics with Qwen API Integration"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.124.0",
    "uvicorn>=0.38.0",
    "fastembed==0.7.4",
    "qdrant-client==1.16.1",
    "pydantic>=2.9.0",
    "python-multipart>=0.0.12",
    "openai>=1.55.0",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.2",
    "asyncpg>=0.31.0",
    "openai-chatkit>=1.4.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["specs*", "history*", "tests*"]
